@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0-windows</TargetFramework>
+    <TargetFramework>net6.0-windows</TargetFramework>
 
     <IsPackable>false</IsPackable>
   </PropertyGroup>
@@ -20,6 +20,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\PSQA.Assets\PSQA.Assets.csproj" />
+    <ProjectReference Include="..\PSQA.Core\PSQA.Core.csproj" />
     <ProjectReference Include="..\PSQuickAssets\PSQuickAssets.csproj" />
   </ItemGroup>
 
