@@ -1,76 +1,123 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-	<OutputType>WinExe</OutputType>
-	<TargetFramework>net5.0-windows</TargetFramework>
-	<UseWPF>true</UseWPF>
-	<ApplicationIcon>Resources\Images\psquickassets-size.ico</ApplicationIcon>
-	<Version>1.1.1</Version>
-	<StartupObject>PSQuickAssets.App</StartupObject>
-	<Authors>Mortuusars</Authors>
-	<PackageProjectUrl>https://github.com/mortuusars/PSQuickAssets</PackageProjectUrl>
-	<RepositoryUrl>https://github.com/mortuusars/PSQuickAssets</RepositoryUrl>
-  </PropertyGroup>
+	<PropertyGroup>
+		<OutputType>WinExe</OutputType>
+		<TargetFramework>net6.0-windows</TargetFramework>
+		<UseWPF>true</UseWPF>
+		<ApplicationIcon>Resources\Images\psquickassets-size.ico</ApplicationIcon>
+		<Version>1.2.0</Version>
+		<StartupObject>PSQuickAssets.App</StartupObject>
+		<Authors>Mortuusars</Authors>
+		<PackageProjectUrl>https://github.com/mortuusars/PSQuickAssets</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/mortuusars/PSQuickAssets</RepositoryUrl>
+		<SourceRevisionId>build$([System.DateTime]::UtcNow.ToString("yyyy-MM-ddTHH:mm:ss:fffZ"))</SourceRevisionId>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-	<None Remove="Resources\Fonts\Lexend-Bold.ttf" />
-	<None Remove="Resources\Fonts\Lexend-ExtraBold.ttf" />
-	<None Remove="Resources\Fonts\Lexend-Light.ttf" />
-	<None Remove="Resources\Fonts\Lexend-Medium.ttf" />
-	<None Remove="Resources\Fonts\Lexend-Regular.ttf" />
-	<None Remove="Resources\Fonts\Lexend-SemiBold.ttf" />
-	<None Remove="Resources\Fonts\Lexend-Thin.ttf" />
-	<None Remove="Resources\Fonts\PTSans-Bold.ttf" />
-	<None Remove="Resources\Fonts\PTSans-BoldItalic.ttf" />
-	<None Remove="Resources\Fonts\PTSans-Regular.ttf" />
-	<None Remove="Resources\Images\close_24.png" />
-	<None Remove="Resources\Images\corner_resize_128.png" />
-	<None Remove="Resources\Images\corner_resize_24.png" />
-	<None Remove="Resources\Images\folder_24.png" />
-	<None Remove="Resources\Images\psb_90.png" />
-	<None Remove="Resources\Images\psd_90.png" />
-	<None Remove="Resources\Images\psquickassets-size.ico" />
-	<None Remove="Resources\Images\settings.png" />
-  </ItemGroup>
+	<ItemGroup>
+		<None Remove="Resources\Fonts\Lexend-Bold.ttf" />
+		<None Remove="Resources\Fonts\Lexend-ExtraBold.ttf" />
+		<None Remove="Resources\Fonts\Lexend-Light.ttf" />
+		<None Remove="Resources\Fonts\Lexend-Medium.ttf" />
+		<None Remove="Resources\Fonts\Lexend-Regular.ttf" />
+		<None Remove="Resources\Fonts\Lexend-SemiBold.ttf" />
+		<None Remove="Resources\Fonts\Lexend-Thin.ttf" />
+		<None Remove="Resources\Fonts\MaterialIcons-Outlined.otf" />
+		<None Remove="Resources\Fonts\PTSans-Bold.ttf" />
+		<None Remove="Resources\Fonts\PTSans-BoldItalic.ttf" />
+		<None Remove="Resources\Fonts\PTSans-Regular.ttf" />
+		<None Remove="Resources\Images\corner_resize_128.png" />
+		<None Remove="Resources\Images\image.png" />
+		<None Remove="Resources\Images\image_dark.png" />
+		<None Remove="Resources\Images\image_light.png" />
+		<None Remove="Resources\Images\psb_90.png" />
+		<None Remove="Resources\Images\psd_90.png" />
+		<None Remove="Resources\Images\psquickassets-size.ico" />
+		<None Remove="Resources\Images\test.jpg" />
+	</ItemGroup>
 
-  <ItemGroup>
-	<PackageReference Include="gong-wpf-dragdrop" Version="2.4.0" />
-	<PackageReference Include="Hardcodet.NotifyIcon.Wpf" Version="1.1.0" />
-	<PackageReference Include="Ookii.Dialogs.Wpf" Version="3.1.0" />
-	  <PackageReference Include="PropertyChanged.Fody" Version="3.4.0">
-		  <PrivateAssets>all</PrivateAssets>
-	  </PackageReference>
-	  <PackageReference Include="SharpHotkeys.WPF" Version="1.0.0" />
-	<PackageReference Include="WpfScreenHelper" Version="1.0.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<None Include="..\.editorconfig" Link=".editorconfig" />
+	</ItemGroup>
 
-  <ItemGroup>
-	<Resource Include="Resources\Fonts\Lexend-Bold.ttf" />
-	<Resource Include="Resources\Fonts\Lexend-ExtraBold.ttf" />
-	<Resource Include="Resources\Fonts\Lexend-Light.ttf" />
-	<Resource Include="Resources\Fonts\Lexend-Medium.ttf" />
-	<Resource Include="Resources\Fonts\Lexend-Regular.ttf" />
-	<Resource Include="Resources\Fonts\Lexend-SemiBold.ttf" />
-	<Resource Include="Resources\Fonts\Lexend-Thin.ttf" />
-	<Resource Include="Resources\Fonts\PTSans-Bold.ttf" />
-	<Resource Include="Resources\Fonts\PTSans-BoldItalic.ttf" />
-	<Resource Include="Resources\Fonts\PTSans-Regular.ttf" />
-	<Resource Include="Resources\Images\close_24.png" />
-	<Resource Include="Resources\Images\corner_resize_24.png" />
-	<Resource Include="Resources\Images\folder_24.png" />
-	<Resource Include="Resources\Images\psb_90.png" />
-	<Resource Include="Resources\Images\psd_90.png" />
-	<Resource Include="Resources\Images\psquickassets-size.ico" />
-	<Resource Include="Resources\Images\settings.png" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="AsyncAwaitBestPractices" Version="6.0.4" />
+		<PackageReference Include="CommunityToolkit.Mvvm" Version="8.0.0-preview3" />
+		<PackageReference Include="gong-wpf-dragdrop" Version="3.1.1" />
+		<PackageReference Include="Hardcodet.NotifyIcon.Wpf" Version="1.1.0" />
+		<PackageReference Include="MGlobalHotkeys.WPF" Version="1.0.2" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
+		<PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.39" />
+		<PackageReference Include="Mortuus.Config" Version="1.0.1" />
+		<PackageReference Include="MTerminal.WPF" Version="0.1.3" />
+		<PackageReference Include="NetVips" Version="2.1.0" />
+		<PackageReference Include="NetVips.Native.win-x64" Version="8.12.2" />
+		<PackageReference Include="Ookii.Dialogs.Wpf" Version="5.0.1" />
+		<PackageReference Include="PureConfig" Version="1.0.0" />
+		<PackageReference Include="PureUI" Version="1.0.18" />
+		<PackageReference Include="PureUI.Themes" Version="1.0.0" />
+		<PackageReference Include="Serilog" Version="2.11.0-dev-01380" />
+		<PackageReference Include="Serilog.Sinks.Debug" Version="2.0.0" />
+		<PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
+		<PackageReference Include="WPFLocalizeExtension" Version="3.9.4" />
+		<PackageReference Include="WpfScreenHelper" Version="2.0.0" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <None Update="Resources\Sounds\click.wav">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-    </None>
-    <None Update="version.json">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\PSQA.Assets\PSQA.Assets.csproj" />
+		<ProjectReference Include="..\PSQA.Core\PSQA.Core.csproj" />
+		<ProjectReference Include="..\PSQuickAssets.AssetControl\PSQuickAssets.AssetControl.csproj" />
+		<ProjectReference Include="..\PSQuickAssets.PSInterop\PSQuickAssets.PSInterop.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<Resource Include="Resources\Fonts\Lexend-Bold.ttf" />
+		<Resource Include="Resources\Fonts\Lexend-ExtraBold.ttf" />
+		<Resource Include="Resources\Fonts\Lexend-Light.ttf" />
+		<Resource Include="Resources\Fonts\Lexend-Medium.ttf" />
+		<Resource Include="Resources\Fonts\Lexend-Regular.ttf" />
+		<Resource Include="Resources\Fonts\Lexend-SemiBold.ttf" />
+		<Resource Include="Resources\Fonts\Lexend-Thin.ttf" />
+		<Resource Include="Resources\Fonts\MaterialIcons-Outlined.otf" />
+		<Resource Include="Resources\Fonts\PTSans-Bold.ttf" />
+		<Resource Include="Resources\Fonts\PTSans-BoldItalic.ttf" />
+		<Resource Include="Resources\Fonts\PTSans-Regular.ttf" />
+		<Resource Include="Resources\Images\image.png" />
+		<Resource Include="Resources\Images\image_dark.png" />
+		<Resource Include="Resources\Images\image_light.png" />
+		<Resource Include="Resources\Images\psb_90.png" />
+		<Resource Include="Resources\Images\psd_90.png" />
+		<Resource Include="Resources\Images\psquickassets-size.ico" />
+		<Resource Include="Resources\Images\test.jpg" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<InternalsVisibleTo Include="PSQuickAssets.Tests"></InternalsVisibleTo>
+	</ItemGroup>
+	
+	<ItemGroup>
+	  <Compile Update="Resources\Lang\Lang.Designer.cs">
+	    <DesignTime>True</DesignTime>
+	    <AutoGen>True</AutoGen>
+	    <DependentUpon>Lang.resx</DependentUpon>
+	  </Compile>
+	  <Compile Update="Windows\AssetsWindow.xaml.cs">
+	    <SubType>Code</SubType>
+	  </Compile>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <EmbeddedResource Update="Resources\Lang\Lang.resx">
+	    <Generator>PublicResXFileCodeGenerator</Generator>
+	    <LastGenOutput>Lang.Designer.cs</LastGenOutput>
+	  </EmbeddedResource>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <Page Update="Windows\AssetsWindow.xaml">
+	    <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
+	    <SubType>Designer</SubType>
+	  </Page>
+	</ItemGroup>
 
 </Project>
