@@ -1,6 +1,4 @@
-﻿using System;
-using System.Globalization;
-using System.Windows.Data;
+﻿using System.Globalization;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 
